third_party/WebKit/LayoutTests/fast/dom/custom/element-type-expected.txt

CONSOLE WARNING: line 28: document.registerElement is deprecated and will be removed in M73, around March 2019. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 for more details.
Test that changing an instantiated custom element's 'is' attribute does not affect its element type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.testRunner is non-null.
PASS a.getAttribute("is") is "x-y"
PASS a instanceof X is true
PASS createdBy is "X"
PASS createdCallCount is 1
PASS successfullyParsed is true

TEST COMPLETE
